requirements.txt
```
PyGithub
requests
transformers
torch
sentencepiece
click
streamlit
python-dotenv
```